requirements.txt
python-telegram-bot==20.7
Flask==2.3.2
requests==2.31.0
APScheduler==3.10.4
